Split by files and write tests
Divided into files and wrote tests

Since validation was removed from the extractor and moved to middleware
and another extractor was added, it was necessary to add tests
