+++ Cargo.toml
@@ -17,10 +17,8 @@ version = "1.0.0"
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
-actix-rt = "2.7.0"
 actix-web = { version="4.2.1", features=["rustls"]}
 biscuit = "0.6.0-beta1"
-bytes = "1.3.0"
 futures = "0.3.25"
 futures-core = {version = "0.3.25", default-features = false, features = ["alloc"]}
 futures-util = {version = "0.3.25", default-features = false, features = ["alloc"]}
@@ -32,3 +30,7 @@ thiserror = "1.0.38"
 
 [dev-dependencies]
 pretty_assertions= "1.3.0"
+actix-rt = "2.7.0"
+bytes = "1.3.0"
+ring = "~0.16.20"
+num = "0.3"
